astronomer_airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google
google-cloud-storage
google-cloud-resource-manager>=1.11.0
apache-airflow-providers-google>=18.0.0
pandas
Faker>=25,<26
